Feat: Chatbot improvements
