+++ models/chat.ts
@@ -2,7 +2,7 @@ type User = "user" | "assistant" | "system"
 
 export interface Message {
   id: string;
-  created_at: string;
+  created_at?: string;
   role: User;
   content: string;
 }
