+++ styles/globals.css
@@ -2,22 +2,57 @@
 @tailwind components;
 @tailwind utilities;
 
-.textarea::-webkit-scrollbar {
+/* Ant Design Modal */
+.ant-modal-custom .ant-modal-content {
+  color: white;
+  background: #070e1a;
+}
+
+.ant-modal-custom .ant-modal-header {
+  color: white;
+  background: transparent;
+}
+
+.ant-modal-custom .ant-modal-title {
+  color: white;
+  font-size: 24px;
+}
+
+.ant-modal-custom .ant-modal-close-x svg {
+  fill: white;
+}
+
+.ant-modal-custom input {
+  margin-top: 32px;
+}
+
+.ant-modal-custom input:focus {
+  outline: none;
+  border: transparent;
+}
+
+.ant-modal-custom .ant-modal-footer {
+  display: flex;
+  justify-content: flex-end;
+}
+
+/* Custom scroll designs */
+.custom-scroll::-webkit-scrollbar {
   width: 4px;
 }
 
 /* Track */
-.textarea::-webkit-scrollbar-track {
+.custom-scroll::-webkit-scrollbar-track {
   background-color: transparent;
 }
 
 /* Handle */
-.textarea::-webkit-scrollbar-thumb {
+.custom-scroll::-webkit-scrollbar-thumb {
   background-color: #32c4a2;
   border-radius: 10px;
 }
 
 /* Handle on hover */
-.textarea::-webkit-scrollbar-thumb:hover {
+.custom-scroll::-webkit-scrollbar-thumb:hover {
   background-color: darkred;
 }
